Added options to CLI, Created TASK component
Added configuration options to the E2ForLife CLI (v1.2) that allow for
more flexible configuration of the component when using it for other
functions such as protocol servers, and non-traditional CLI type
applications.  Also created a sandbox task component to add FreeRTOS
tasks without having to write the other task code, create the task and
mutexes.  Basically the task component allows us to be lazy and reuse
code to create tasks, or to create task-based drivers that can be cloned
to components later.